feat: add DependencyInjection project

+++ BarberShop/BarberShop.Domain/Interfaces/IBookingRepository.cs
@@ -4,7 +4,7 @@ namespace BarberShop.Domain.Interfaces;
 
 public interface IBookingRepository
 {
-    Task<Booking> Get(int id);
-    Task<IEnumerable<Booking>> GetByUserId(int userId);
+    Task<Booking> Get(long id);
+    Task<IEnumerable<Booking>> GetByUserId(long userId);
     Task<IEnumerable<Booking>> GetAll();
 }
